# backend/server.js
require('dotenv').config();
const express = require('express');
const bodyParser = require('body-parser');
const axios = require('axios');
const authenticate = require('./src/authenticate');

const app = express();
const port = process.env.PORT || 8080;

app.use(bodyParser.json());

async function getAccessToken() {
    const tokenUrl = process.env.CHAT_SERVICE_OAUTH_TOKEN_URL;
    const clientId = process.env.CHAT_SERVICE_OAUTH_CLIENT_ID;
    const clientSecret = process.env.CHAT_SERVICE_OAUTH_CLIENT_SECRET;

    try {
        const accessToken = await authenticate(tokenUrl, clientId, clientSecret);
        return accessToken;
    } catch (error) {
        console.error('Error obtaining access token:', error);
        throw error; 
    }
}

app.get('/ping', async (req, res) => {
    const accessToken = await getAccessToken(); 
    const chatServiceUrl = process.env.CHAT_SERVICE_URL;
    const response = await axios.get(`${chatServiceUrl}/ping`, {
        headers: {
            'Authorization': `Bearer ${accessToken}`,
        },
    });
    res.status(response.status).send(response.data);
});

app.post('/auth/login', async (req, res) => {
    try {
        const { username, password } = req.body;
        const accessToken = await getAccessToken(); 

        const chatServiceUrl = process.env.CHAT_SERVICE_URL;
        const response = await axios.post(`${chatServiceUrl}/login`, 
        {
            username,
            password,
        },
        req.body, {
            headers: {
                'Authorization': `Bearer ${accessToken}`,
            },
        });
        res.status(response.status).send(response.data);
    } catch (error) {
        console.error('Error login:', error);
        res.status(error.response ? error.response.status : 500).send(error.message);
    }
});

app.post('/auth/register', async (req, res) => {
    try {
        const { username, email, password } = req.body;
        const accessToken = await getAccessToken(); 
        const chatServiceUrl = process.env.CHAT_SERVICE_URL;
        const response = await axios.post(`${chatServiceUrl}/register`, 
        {
            username,
            email,
            password
        },
        req.body, {
            headers: {
                'Authorization': `Bearer ${accessToken}`,
            },
        });
        res.status(response.status).send(response.data);
    } catch (error) {
        console.error('Error register:', error);
        res.status(error.response ? error.response.status : 500).send(error.message);
    }
});

app.get('/auth/allusers', async (req, res) => {
    try {
        const id = req.query.id;
        const accessToken = await getAccessToken(); 

        const chatServiceUrl = process.env.CHAT_SERVICE_URL;
        const response = await axios.get(`${chatServiceUrl}/allusers=${id}`, 
        req.body, {
            headers: {
                'Authorization': `Bearer ${accessToken}`,
            },
        });
        res.status(response.status).send(response.data);
    } catch (error) {
        console.error('Error getting user:', error);
        res.status(error.response ? error.response.status : 500).send(error.message);
    }
});


app.get('/auth/logout', async (req, res) => {
    try {
        const id = req.query.id;
        const accessToken = await getAccessToken(); 

        const chatServiceUrl = process.env.CHAT_SERVICE_URL;
        const response = await axios.get(`${chatServiceUrl}/logout=${id}`, 
        req.body, {
            headers: {
                'Authorization': `Bearer ${accessToken}`,
            },
        });
        res.status(response.status).send(response.data);
    } catch (error) {
        console.error('Error logging out:', error);
        res.status(error.response ? error.response.status : 500).send(error.message);
    }
});

app.get('/auth/setAvatar', async (req, res) => {
    try {
        const id = req.query.id;
        const accessToken = await getAccessToken(); 

        const chatServiceUrl = process.env.CHAT_SERVICE_URL;
        const response = await axios.post(`${chatServiceUrl}/setAvatar=${id}`, 
        req.body, {
            headers: {
                'Authorization': `Bearer ${accessToken}`,
            },
        });
        res.status(response.status).send(response.data);
    } catch (error) {
        console.error('Error set avatar:', error);
        res.status(error.response ? error.response.status : 500).send(error.message);
    }
});

app.listen(port, () => {
    console.log(`Server running on port ${port}`);
});